--- demo.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-
-    <script src="index.js"></script>
-    <script src="renderArea.js"></script>
-    <style>
-        html,body{width:100%;height:100%;margin:0;padding:0;}
-    </style>
-</head>
-<body>
-<canvas id="cvs" style="margin:0;padding: 0;width:100%;height:100%;">
-    cvs
-</canvas>
-<script>
-    var ctx = document.getElementById("cvs").getContext("2d");
-    var canvasSize = {top:0,left:0,width: window.innerWidth,height:window.innerHeight};
-    ctx.canvas.width  = window.innerWidth;
-    ctx.canvas.height = window.innerHeight;
-    window.onresize=function(){
-        ctx.canvas.width  = window.innerWidth;
-        ctx.canvas.height = window.innerHeight;
-        canvasSize = {top:0,left:0,width: window.innerWidth,height:window.innerHeight};
-        console.log("onresize");
-    };
-    document.onkeyup = keyPressEvent;
-    document.getElementById("cvs").onwheel = wheelEvent;
-    document.getElementById("cvs").onmousemove = mouseMoveEvent;
-    document.getElementById("cvs").onmouseup = mouseReleaseEvent;
-    document.getElementById("cvs").onmousedown = mousePressEvent;
-    document.getElementById("cvs").oncontextmenu=function(){return false;};
-    var paint =paintEvent;
-    initPolygons();
-    paint();
-
-</script>
-</body>
-</html>
