--- index.js
@@ -1,6 +1,0 @@
-module.exports = {
-    Polygon : require('./src/polygon.js'),
-    Vector: require('./src/vector.js').Vector,
-    Vectors: require('./src/vector.js').Vectors,
-    Line: require('./src/line.js')
-}
