+++ js/main.js
@@ -1,0 +1,12 @@
+const Vector = require('./vector').default;
+const Vectors = require('./vectors').default;
+const Line = require('./line').default;
+const Polygon = require('./polygon').default;
+const Polygons = require('./polygons').default;
+
+
+module.exports = Vector;
+module.exports = Vectors;
+module.exports = Line;
+module.exports = Polygon;
+module.exports = Polygons;
