--- test.js
@@ -1,3 +1,0 @@
-// const Polygon = require('./src/polygon.js');
-const Polygon = require('./index.js');
-console.log(Polygon);
