+++ webpack.config.js
@@ -1,0 +1,42 @@
+const path = require('path');
+const webpack = require('webpack');
+
+var CopyWebpackPlugin = require('copy-webpack-plugin');
+
+var dir_js = path.resolve(__dirname, 'js');
+var dir_html = path.resolve(__dirname, 'html');
+var dir_build = path.resolve(__dirname, 'build');
+
+module.exports = {
+    entry: path.resolve(dir_js, 'main.js'),
+    output: {
+        path: dir_build,
+        filename: 'bundle.js'
+    },
+    module: {
+        rules: [
+            {
+                test: /\.(js|mjs|jsx|ts|tsx)$/,
+                include: dir_js,
+                exclude: /(node_modules|bower_components)/,
+
+
+                loader: 'babel-loader',
+
+
+            }
+        ]
+
+    },
+    plugins: [
+        // Simply copies the files over
+        new CopyWebpackPlugin([
+            { from: dir_html } // to: output.path
+        ]),
+    ],
+    // Create Sourcemaps for the bundle
+    devtool: 'source-map',
+    devServer: {
+        contentBase: dir_build,
+    },
+};
